281 (#41)
- improve `h5tbx serve` RDF browser
  - add `--recursive` for recursive folder discovery
  - add `--include-ttl` to include Turtle RDF files in the combined graph
  - improve graph layout with collapsible settings and a larger usable graph view
- refactor `h5tbx` CLI to Typer for improved help output and command-line ergonomics while preserving existing commands
- improve SHACL validation with ontology graph support
  - add `ont_graph`, `ont_graph_format`, and `merge_ont_graph_into_data` to `validate_hdf(...)`
  - use `hdf_data_format` when parsing string RDF data passed to `validate_hdf(...)`
- improve JSON-LD import handling before converting JSON-LD metadata to HDF5
  - resolve `@import` contexts through the toolbox context cache before RDFLib processing
  - update metadata4ing/m4i context URL handling, normalize legacy URLs, and provide an offline fallback context
- improve RDF graph labels by using compact Zenodo prefixes `zen:` and `rzen:`
- update FAQ, metadata4ing, and SHACL validation documentation

+++ tests/database/test_mongo.py
@@ -110,7 +110,7 @@ class TestH5Mongo(unittest.TestCase):
 
         with h5py.File("test.h5", "w") as h5:
             h5.create_dataset("dataset0d", data=1)
-            h5.create_dataset("dataset", shape=(10, 20, 4))
+            h5.create_dataset("dataset", shape=(10, 20, 4), dtype="f4")
 
             mongoDBInterface.insert_dataset(h5["dataset0d"], axis=None)
             mongoDBInterface.insert_dataset(h5["dataset"], axis=None)
@@ -137,7 +137,7 @@ class TestH5Mongo(unittest.TestCase):
         self.client.drop_database("hdf_database_test")
         mongoDBInterface = MongoDB(collection=self.collection)
         with h5py.File("test.h5", "w") as h5:
-            h5.create_dataset("dataset", shape=(10, 20, 4))
+            h5.create_dataset("dataset", shape=(10, 20, 4), dtype="f4")
 
             mongoDBInterface.insert_dataset(h5["dataset"], axis=0)
         self.assertEqual(10, mongoDBInterface.collection.count_documents({}))
@@ -160,7 +160,7 @@ class TestH5Mongo(unittest.TestCase):
 
         with h5py.File("test.h5", "w") as h5:
             h5.create_dataset("grp/dataset0d", data=1)
-            h5.create_dataset("grp/subgrp/dataset", shape=(10, 20, 4))
+            h5.create_dataset("grp/subgrp/dataset", shape=(10, 20, 4), dtype="f4")
 
             mongoDBInterface.insert_group(h5["grp"], recursive=False)
 
@@ -171,7 +171,7 @@ class TestH5Mongo(unittest.TestCase):
         self.assertEqual(0, mongoDBInterface.collection.count_documents({}))
 
         with h5py.File("test.h5", "w") as h5:
-            h5.create_dataset("dataset", shape=(10, 20, 4))
+            h5.create_dataset("dataset", shape=(10, 20, 4), dtype="f4")
 
             mongoDBInterface.insert_dataset(h5["dataset"], axis=0)
